requirements.txt
# Core LLM / Agent
langchain-mistralai>=0.2.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-text-splitters>=0.3.0
langchain-huggingface>=0.1.0
langgraph>=0.2.0
faiss-cpu>=1.8.0
sentence-transformers>=3.0.0

# Web scraping for RAG
beautifulsoup4>=4.12.0
requests>=2.32.0

# API server
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.7.0
python-dotenv>=1.0.0
